# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Billing API
  version: 0.2.0
  description: Subscription billing — products, prices, coupons, subscriptions, invoices, and credit notes
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3002
    description: Local billing service. Override via Postman {{baseUrl}} environment variable.
...
paths:
...
  /api/v1/promotion_codes:
    get:
      summary: List promotion codes
      description: Returns a paginated list of promotion codes with optional coupon and active filters.
      x-platform:
        kind: resource
        entity: PromotionCode
        affectsLedger: false
      operationId: listPromotionCodes
      tags:
        - PromoCodes
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - in: query
          name: coupon_id
          description: Filter by coupon ID.
          schema:
            type: string
        - in: query
          name: active
          description: Filter by active status.
          schema:
            type: string
            enum:
              - "true"
              - "false"
      responses:
        "200":
          description: Paginated list of promotion codes.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create promotion code
      description: Creates a promotion code linked to an active coupon. Code must be unique per merchant.
      x-platform:
        kind: resource
        entity: PromotionCode
        emitsEvents:
          - promotion_code.created
        affectsLedger: false
      operationId: createPromotionCode
      tags:
        - PromoCodes
      security:
        - MerchantKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - coupon_id
                - code
              properties:
                coupon_id:
                  type: string
                  description: ID of the coupon this promotion code references.
                code:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: Unique code string per merchant.
                max_redemptions:
                  type: integer
                  minimum: 1
                expires_at:
                  type: string
                  format: date-time
                restrictions:
                  type: object
                  additionalProperties: true
                metadata:
                  type: object
                  additionalProperties: true
            examples:
              basic:
                value:
                  coupon_id: cpn_abc123
                  code: SUMMER2026
                  max_redemptions: 100
      responses:
        "201":
          description: Promotion code created.
          content:
            application/json:
              schema:
                type: object
                required:
                  - promotion_code
                properties:
                  promotion_code:
                    type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Coupon not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Code already exists or coupon is inactive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/subscriptions/{id}/items/{itemId}:
    get:
      summary: Get subscription item
      description: Retrieves a single subscription item.
      x-platform:
        kind: resource
        entity: SubscriptionItem
        affectsLedger: false
      operationId: getSubscriptionItem
      tags:
        - Subscriptions
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - in: path
          name: itemId
          required: true
          schema:
            type: string
          description: Subscription item ID.
      responses:
        "200":
          description: Subscription item details.
          content:
            application/json:
              schema:
                type: object
                required:
                  - subscription_item
                properties:
                  subscription_item:
                    $ref: "#/components/schemas/SubscriptionItem"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update subscription item quantity
      description: Updates the quantity of a subscription item.
      x-platform:
        kind: resource
        entity: SubscriptionItem
        emitsEvents:
          - subscription.updated
        affectsLedger: false
      operationId: updateSubscriptionItem
      tags:
        - Subscriptions
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - in: path
          name: itemId
          required: true
          schema:
            type: string
          description: Subscription item ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: Item updated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - subscription_item
                properties:
                  subscription_item:
                    $ref: "#/components/schemas/SubscriptionItem"
        "400":
          description: Subscription terminated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Remove subscription item
      description: Removes a subscription item. Cannot remove the last item.
      x-platform:
        kind: transition
        entity: SubscriptionItem
        emitsEvents:
          - subscription.updated
        affectsLedger: false
        stateChanges:
          - item_removed
      operationId: deleteSubscriptionItem
      tags:
        - Subscriptions
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - in: path
          name: itemId
          required: true
          schema:
            type: string
          description: Subscription item ID.
      responses:
        "200":
          description: Item removed.
          content:
            application/json:
              schema:
                type: object
                required:
                  - deleted
                properties:
                  deleted:
                    type: boolean
        "400":
          description: Cannot remove last item or subscription terminated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
    SubscriptionItem:
      description: A line item on a subscription linking a price.
      type: object
      properties:
        id:
          type: string
        subscription_id:
          type: string
        price_id:
          type: string
        quantity:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  parameters:
...
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...